simple_django/forms/requirements.txt
amqp==5.0.1
appdirs==1.4.3
asgiref==3.2.10
attrs==20.3.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
billiard==3.6.3.0
CacheControl==0.12.6
cached-property==1.5.2
celery==5.0.2
certifi==2019.11.28
cffi==1.14.4
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
colorama==0.4.3
contextlib2==0.6.0
cryptography==3.2.1
distlib==0.3.0
distro==1.5.0
Django==3.1.2
django-bootstrap-form==3.4
django-debug-toolbar==3.1.1
django-rest-framework==0.1.0
django-suit==0.2.25
djangorestframework==3.12.2
docker==4.4.0
docker-compose==1.27.4
dockerpty==0.4.1
docopt==0.6.2
Faker==4.17.1
flake8==3.8.4
flake8-builtins==1.5.3
flake8-django==1.1.1
flake8-import-order==0.6
flake8-print==3.1.4
Flask==1.1.2
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
kombu==5.0.2
lockfile==0.12.2
lxml==4.6.1
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==0.6.2
packaging==20.3
paramiko==2.7.2
pep517==0.8.2
pep8==1.7.1
progress==1.5
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PyNaCl==1.4.0
pyparsing==2.4.6
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.15.0
pytoml==0.1.21
pytz==2020.1
PyYAML==5.3.1
randstr==0.1.2
requests==2.22.0
retrying==1.3.3
six==1.14.0
soupsieve==2.0.1
sqlparse==0.4.1
text-unidecode==1.3
texttable==1.6.3
urllib3==1.25.8
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
yum==0.0.1
